make image fields all floatfields

+++ server/annotations/models.py
@@ -7,10 +7,10 @@ class Annotation(models.Model):
     """
 
     type = models.CharField(max_length=100)
-    startX = models.IntegerField()
-    startY = models.IntegerField()
-    endX = models.IntegerField()
-    endY = models.IntegerField()
+    startX = models.FloatField()
+    startY = models.FloatField()
+    endX = models.FloatField()
+    endY = models.FloatField()
     image = models.ForeignKey(
         "Image", related_name="annotations", on_delete=models.CASCADE, null=True
     )
@@ -22,8 +22,8 @@ class Image(models.Model):
         "Project", related_name="images", on_delete=models.CASCADE, null=True
     )
     name = models.CharField(max_length=100)
-    width = models.IntegerField()
-    height = models.IntegerField()
+    width = models.FloatField()
+    height = models.FloatField()
 
 
 class Project(models.Model):
